openstack error handling is inside openstack module
